requirements.txt
future
django==1.11.2
jinja2
djangorestframework
